Move height tests
